changed amount to double

+++ src/main/java/supermarket/Article.java
@@ -10,7 +10,7 @@ public class Article {
 		this.price = price;
 	}
 
-	public double total(int amount) {
+	public double total(double amount) {
 		return price.total(amount);
 	}
